Remove Delegated and its pattern
The idea behind Delegated is to abstract away not capturing self in a block, and thus avoiding retain cycles.
But, it was doing so at the expense of lots of duplicated code (e.g. onClose() and delegatedOnClose(), onError() and delegatedOnError(), and so on).
It also requires blocks with multiple arguments to pass them as a tuple that then has to be decomposed within the block. Once adopted it ends up being used everywhere, even in blocks that don't want to capture self.

On the other hand, without it, callers must be careful not to capture self in escaping blocks. This is standard in Swift code and many other Swift libraries, as well as Foundation, UIKit, etc. require awareness of retain cycles.