Remove JSONElement from the project; keep message payloads as Data
- Update PayloadEncoder/PayloadDecoder and dismabiguate coding Any vs Codable with different method names rather than matching an overload
- Use PayloadEncoder/PayloadDecoder to roundtrip message payload back to Data before passing on as a Message

The roundtrip is a disadvantage over 5.x, which just used JSONSerialization and provided a [String: Any] of the payload.
But, the other elements of the incoming message have to be parsed to know how to handle the payload, and Phoenix Channels don't provide a wrapper around the payload for non-binary channel messages. So, if the payload needs to come out as Data (or as a Codable type decoded from that data), some sort of partial decoding or roundtripping is necessary.

JSONSerialization is almost certainly the fastest and most robust way to do that.